# tests/MurciaDatos.Api.Tests/Contrato/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "API de datos de turismo de Murcia",
    "description": "Demanda, oferta y precios del alojamiento turístico en la Región de Murcia (INE y murciaturistica.es), servidos desde la\r\nrelease mensual del proyecto murcia-open-data.\r\n\r\n**Convenciones**\r\n- Solo lectura (`GET`). Prefijo `/v1`. Identificadores en `kebab-case`; fechas como `AAAA-MM`.\r\n- Un dato que no existe (un mes sin publicar, una medida que no aplica) es `null`, **nunca 0**.\r\n- Los valores provisionales se marcan con `provisional`; `meta.provisional_desde` dice desde qué mes lo son.\r\n- Los errores son `application/problem+json` (RFC 9457) y enumeran los valores permitidos.\r\n- JSON o CSV (`formato=csv` o `Accept: text/csv`; `excel=true` para abrirlo directamente en Excel en español).\r\n- Caché: `ETag` y `If-None-Match` (304). Límite de peticiones por IP: cabeceras `RateLimit-*`; al pasarse, 429 con `Retry-After`.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/demanda": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Viajeros y pernoctaciones",
        "description": "Demanda de alojamiento turístico por mes, trimestre o año. Los flujos se **suman** al agregar. Un mes sin dato es `null`, nunca 0; con `agregacion` distinta de `mes`, `meses` dice cuántos meses con dato entran en cada periodo (un año en curso tiene menos de 12).",
        "operationId": "ConsultarDemanda",
        "parameters": [
          {
            "name": "territorio",
            "in": "query",
            "description": "Identificadores de territorio separados por comas (consulta /v1/territorios). Sin indicar: todos.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "description": "Tipos de alojamiento separados por comas: hotel, apartamento, camping, rural. Sin indicar: todos los que haya.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "residencia",
            "in": "query",
            "description": "Solo demanda. «espana», «extranjero», ambos separados por coma, o «total» para sumarlos. Sin indicar: una fila por residencia.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "description": "Primer mes, como AAAA-MM.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "description": "Último mes, como AAAA-MM.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agregacion",
            "in": "query",
            "description": "mes (por defecto), trimestre o anio. Los flujos se suman; las existencias, tasas e índices se promedian.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "medidas",
            "in": "query",
            "description": "Medidas separadas por comas. Sin indicar: todas las del recurso.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "description": "json (por defecto) o csv. También se puede pedir con la cabecera Accept: text/csv.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excel",
            "in": "query",
            "description": "Con formato=csv: true para separar con punto y coma y decimales con coma, y añadir BOM UTF-8 (para abrirlo directamente en Excel en español).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaDeSerie"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/oferta": {
      "get": {
        "tags": [
          "Series"
        ],
        "summary": "Establecimientos, plazas, ocupación y empleo",
        "description": "Oferta de alojamiento y su ocupación. Son **existencias y tasas**: al agregar por trimestre o año se hace la **media** de los meses. Las medidas que no aplican a un tipo (las parcelas en un hotel) son `null`.",
        "operationId": "ConsultarOferta",
        "parameters": [
          {
            "name": "territorio",
            "in": "query",
            "description": "Identificadores de territorio separados por comas (consulta /v1/territorios). Sin indicar: todos.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "description": "Tipos de alojamiento separados por comas: hotel, apartamento, camping, rural. Sin indicar: todos los que haya.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "residencia",
            "in": "query",
            "description": "Solo demanda. «espana», «extranjero», ambos separados por coma, o «total» para sumarlos. Sin indicar: una fila por residencia.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "desde",
            "in": "query",
            "description": "Primer mes, como AAAA-MM.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasta",
            "in": "query",
            "description": "Último mes, como AAAA-MM.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "agregacion",
            "in": "query",
            "description": "mes (por defecto), trimestre o anio. Los flujos se suman; las existencias, tasas e índices se promedian.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "medidas",
            "in": "query",
            "description": "Medidas separadas por comas. Sin indicar: todas las del recurso.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "formato",
            "in": "query",
            "description": "json (por defecto) o csv. También se puede pedir con la cabecera Accept: text/csv.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "excel",
            "in": "query",
            "description": "Con formato=csv: true para separar con punto y coma y decimales con coma, y añadir BOM UTF-8 (para abrirlo directamente en Excel en español).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespuestaDeSerie"
                }
              },
              "text/csv": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "304": {
            "description": "Not Modified"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MetaDeSerie": {
        "required": [
          "version_datos",
          "fuentes",
          "provisional_desde",
          "agregacion",
          "medidas",
          "total_filas"
        ],
        "type": "object",
        "properties": {
          "version_datos": {
            "type": "string"
          },
          "fuentes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "provisional_desde": {
            "type": [
              "null",
              "string"
            ]
          },
          "agregacion": {
            "type": "string"
          },
          "medidas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MedidaPublica"
            }
          },
          "total_filas": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RespuestaDeSerie": {
        "required": [
          "datos",
          "meta"
        ],
        "type": "object",
        "properties": {
          "datos": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/MetaDeSerie"
          }
        },
        "description": "Respuesta de las series. Cada fila de `datos` tiene las claves `periodo`, `territorio`, `tipo`,\r\n(`residencia` en la demanda), `fuente`, `meses` (si se agrega), `provisional` y una clave por medida pedida.\r\nUn dato que no existe es `null`, nunca 0."
      },
...
    }
  },
...
}